pyproject.toml
[project]
name = "RizzAI"
version = "0.1.0"
description = "A Rizzler bot that uses AI to help you flirt."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { file = "LICENSE" }
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "datasets>=2.14.0",
    "dotenv>=0.9.9",
    "geckodriver-autoinstaller>=0.1.0",
    "peft>=0.7.0",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "trl>=0.7.0",
    "types-requests>=2.32.4.20250913",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.1",
    "typer>=0.19.2",
]
